Update pinout for zigbit module
This updates the pinout for the "ATmega256RFR2 ZigBit Xplained Pro
Extension" (ATZB-256RFR2-XPRO-ND). It uses the green and yellow leds to
indicate flashing progress, and the red led to indicate an invalid boot
(no MCUSR bit set).

Additionally, it uses the first 4 pins of J3 (PF0 - PF3) for logic
analyzer output.