Add compiled 16u2 firmware and bootloader
These are the versions that are used on the Pinoccio Scout v1.0 boards
that have been shipped.

The source of these .hex files is a bit fuzzy. It's not known what
compiler versions were used exactly (the bootloader has been compiled
before the makefile existed, so it was even compiled by two different
compiler versions, one in Atmel Studio for the bootloader and one
through a Makefile for Uracoli).

For this reason, there are also no .elf or .map files available.

However, the source files that are in this git revision should
correspond to the sources used to compile these .hex files (except for
some debug-changes in the bootloader that are not enabled in this
build).