bootloader: Simplify spm_wait slightly
This removes the boot_spm_busy check, which isn't really needed.
spm_wait is always executed just after the start of an spm operation, so
there shouldn't be any chance of it being done already.

However, even if it does happen, the SPM interrupt seems to be a
level-interrupt, so it continuously fires when enabled and no SPM
instruction is currently busy. So at worst, we have one extra sleep
cycle when this unlikely case happens, but we save a check in any case.

Because SPM is a level interrupt, the cli() instruction is moved up to
directly after the sleep. I believe the previous code would execute the
ISR twice every time because of this.

For consistency, this also disables the SPMIE bit again after setting
it, which might prevent confusion in the user program if it assumes the
default value of 0.

Finally, this uses some functions from avr/sleep.h instead of manually
touching the sleep registers. The effective code generated should be the
same.